requirements.txt
gspread==5.12.4
oauth2client==4.1.3
pandas==2.2.1
matplotlib==3.8.3
requests==2.31.0
streamlit==1.32.0
plotly==5.19.0
python-dotenv==1.0.1 